refactor gradio
